+++ Intersect (Core)/GameObjects/SpellBase.cs
@@ -156,8 +156,8 @@ namespace Intersect.GameObjects
     public class SpellWarpData
     {
         public Guid MapId { get; set; }
-        public int X { get; set; }
-        public int Y { get; set; }
+        public byte X { get; set; }
+        public byte Y { get; set; }
         public byte Dir { get; set; }
     }
